Fix multiple warnings (in the context of v2.4.0) (#1285)
* Fix some unused variable warnings

* Fix a user-defined MPI Comm test with unused Comm

* Fix a few reorder warnings

* Fix a few sign-compare warnings

* Format code

+++ tests/parallel/direct-mesh-access/helpers.hpp
@@ -12,6 +12,6 @@ void runTestAccessReceivedMesh(const TestContext &       context,
                                const std::vector<double> writeDataSecondaryRank,
                                const std::vector<double> expectedPositionSecondaryRank,
                                const std::vector<double> expectedReadDataSecondaryRank,
-                               const int                 startIndex);
+                               const size_t              startIndex);
 
 #endif
